+++ src/main/java/com/becoder/service/impl/UserServiceImpl.java
@@ -1,134 +1,40 @@
 package com.becoder.service.impl;
 
-import java.util.List;
-import java.util.UUID;
-
-import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
-import org.springframework.util.ObjectUtils;
 
-import com.becoder.config.security.CustomUserDetails;
-import com.becoder.dto.EmailRequest;
-import com.becoder.dto.LoginRequest;
-import com.becoder.dto.LoginResponse;
-import com.becoder.dto.UserRequest;
-import com.becoder.entity.AccountStatus;
-import com.becoder.entity.Role;
+import com.becoder.dto.PasswordChangRequest;
 import com.becoder.entity.User;
-import com.becoder.repository.RoleRepository;
 import com.becoder.repository.UserRepository;
-import com.becoder.service.JwtService;
 import com.becoder.service.UserService;
-import com.becoder.util.Validation;
+import com.becoder.util.CommonUtils;
 
 @Service
 public class UserServiceImpl implements UserService {
 
 	@Autowired
-	private UserRepository userRepo;
+	private BCryptPasswordEncoder passwordEncoder;
 
 	@Autowired
-	private ModelMapper mapper;
-
-	@Autowired
-	private RoleRepository roleRepo;
-
-	@Autowired
-	private Validation validation;
-
-	@Autowired
-	private EmailService emailService;
-
-	@Autowired
-	private AuthenticationManager authenticationManager;
-
-	@Autowired
-	private PasswordEncoder passwordEncoder;
-
-	@Autowired
-	private JwtService jwtService;
+	private UserRepository userRepository;
 
 	@Override
-	public boolean register(UserRequest userDto, String url) throws Exception {
+	public void changePassword(PasswordChangRequest passChangRequest) {
 
-		validation.userValidation(userDto);
+		User loggedInUser = CommonUtils.getLoggedInUser();
 
-		User user = mapper.map(userDto, User.class);
+		if (!passwordEncoder.matches(passChangRequest.getOldPassword(), loggedInUser.getPassword())) {
 
-		setRole(userDto, user);
-
-		AccountStatus status = AccountStatus.builder().isActive(false).verificationCode(UUID.randomUUID().toString())
-				.build();
-
-		user.setStatus(status);
-		user.setPassword(passwordEncoder.encode(user.getPassword()));
-
-		User saveUser = userRepo.save(user);
-
-		if (!ObjectUtils.isEmpty(saveUser)) {
-
-			emailSend(saveUser, url);
-
-			return true;
+			throw new IllegalArgumentException("Old password is invalid");
 		}
 
-		return false;
-	}
+		String encodePassword = passwordEncoder.encode(passChangRequest.getNewPassword());
 
-	private void emailSend(User saveUser, String url) throws Exception {
+		loggedInUser.setPassword(encodePassword);
 
-		String message = "Hi,<b>  [[username]]  </b>" + "<br> Your Account Register Successfully.<br>"
-				+ "<br> Click the below link verify & Active your account <br>"
-				+ "<a href='[[url]]'>Click Here</a> <br><br>" + "Thanks,<br>Enotes.com";
+		userRepository.save(loggedInUser);
 
-		message = message.replace("[[username]]", saveUser.getFirstName());
-		message = message.replace("[[url]]", url + "/api/v1/home/verify?uid=" + saveUser.getId() + "&&code="
-				+ saveUser.getStatus().getVerificationCode());
-
-		EmailRequest emailRequest = EmailRequest.builder().to(saveUser.getEmail())
-				.title("Account Creating Confirmation").subject("Account Created Success").message(message).build();
-
-		emailService.senEmail(emailRequest);
-
-	}
-
-	private void setRole(UserRequest userDto, User user) {
-
-		List<Integer> reqRoleId = userDto.getRoles().stream().map(r -> r.getId()).toList();
-
-		List<Role> roles = roleRepo.findAllById(reqRoleId);
-
-		user.setRoles(roles);
-	}
-
-	@Override
-	public LoginResponse login(LoginRequest loginRequest) {
-
-		Authentication authenticate = authenticationManager.authenticate(
-
-				new UsernamePasswordAuthenticationToken(loginRequest.getEmail(), loginRequest.getPassword()));
-
-		if (authenticate.isAuthenticated()) {
-
-			CustomUserDetails customUserDetails = (CustomUserDetails) authenticate.getPrincipal();
-
-			String token = jwtService.generateToken(customUserDetails.getUser());
-
-			LoginResponse loginResponse = LoginResponse.builder()
-					.user(mapper.map(customUserDetails.getUser(), UserRequest.class)).token(token)
-
-					.build();
-
-			return loginResponse;
-
-		}
-
-		return null;
 	}
 
 }
